Split the interrupt driver from interrupt policy
gic now only knows registers. The daifclr moved out to a new irq module,
so interrupts go live in kmain as a phase of its own rather than as a
side effect of bringing the controller up - which leaves room for device
drivers to register handlers in between.

Acknowledging returns an Interrupt value instead of a raw word. It is
must_use, end() consumes it, and spurious comes back as None, so
double-EOI, forgetting to EOI, and EOI-ing a spurious interrupt all stop
compiling instead of hanging the controller in silence.

Base addresses collected in board.rs, and mem takes usize now that every
address flows from there.

+++ src/mem.rs
@@ -1,17 +1,17 @@
 use core::ptr::{read_volatile, write_volatile};
 
-pub fn read_byte(address: u64) -> u8 {
+pub fn read_byte(address: usize) -> u8 {
     unsafe { read_volatile(address as *const u8) }
 }
 
-pub fn read_32(address: u64) -> u32 {
+pub fn read_32(address: usize) -> u32 {
     unsafe { read_volatile(address as *const u32) }
 }
 
-pub fn write_byte(address: u64, byte: u8) -> () {
+pub fn write_byte(address: usize, byte: u8) -> () {
     unsafe { write_volatile(address as *mut u8, byte) };
 }
 
-pub fn write_32(address: u64, byte: u32) -> () {
+pub fn write_32(address: usize, byte: u32) -> () {
     unsafe { write_volatile(address as *mut u32, byte) };
 }
